pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel", "pybind11>=2.11"]
build-backend = "setuptools.build_meta"

[project]
name = "turbine-blade-moo"
version = "0.1.0"
description = "From-scratch wind turbine blade optimization using XFOIL + BEM + NSGA-II."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "numpy>=1.24",
  "pandas>=2.0",
  "matplotlib>=3.7",
  "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "jupyter>=1.0",
  "ipykernel>=6.0",
]

[project.scripts]
tb-moo = "turbine_blade_moo.pipeline:cli"

[tool.setuptools.packages.find]
include = ["turbine_blade_moo*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
